pyproject.toml
[project]
name = "tl-2025"
version = "0.1.0"
description = "Chatbot con Chainlit y Google Gemini para TalentLand 2025"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.4.1",
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.1",
    "pyjwt>=2.8.0",
    "argparse>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.1.0",
    "ruff>=0.0.52",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 88
target-version = "py311"
